First_Agent/pyproject.toml
[project]
name = "first-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.3.0",
    "python-dotenv>=1.1.1",
]
